% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{discF2}
\alias{discF2}
\title{Volume of the overlapping region}
\usage{
discF2(data, class.idx = ncol(data))
}
\arguments{
\item{data}{A data frame with discrete variables.}

\item{class.idx}{An optional numerical index for the position of the class
variable in the data frame. It is assumed the last variable in the data.}
}
\value{
A float number representing the volume of the overlapping region.
}
\description{
Compute the ratio of overlapped examples in a discrete data frame.
}
\examples{
data(qb)
F2 <- discF2(qb)
F2
}
