% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{discF4}
\alias{discF4}
\title{Collective Feature Efficiency}
\usage{
discF4(data, class.idx = ncol(data))
}
\arguments{
\item{data}{A data frame with discrete variables.}

\item{class.idx}{An optional numerical index for the position of the class
variable in the data frame. It is assumed the last variable in the data.}
}
\value{
A list with the value of F4 (value), a nested list of the best
attributes to separate examples (attribs), a nested list of indexes of
examples with overlapping (overlapping) and a nested list of indexes of
examples with no overlapping (nonOverlapping).
}
\description{
Compute the ratio of examples with overlapping in a nominal data frame.
}
\examples{
data(qb)
F4 <- discF4(qb[1:15,])
F4$value
F4$attribs
F4$overlapping
F4$nonOverlapping
}
