% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{generateDF}
\alias{generateDF}
\title{Generation of synthetic data}
\usage{
generateDF(model, size)
}
\arguments{
\item{size}{An integer representing the number of generated examples.}

\item{data}{A data frame with discrete variables.}
}
\value{
A data frame.
}
\description{
Generate a data frame with an specific number of examples according the
distribution of a noise model.
}
\examples{
data(qb)
model <- NCAR(data = qb, noise = 0.45)
df <- generateDF(model$parameters, 10)
df
}
