% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{inconsistency}
\alias{inconsistency}
\title{Percentage of inconsistent examples}
\usage{
inconsistency(data, class.idx = ncol(data))
}
\arguments{
\item{data}{A data frame with discrete variables.}

\item{class.idx}{An optional numerical index for the position of the class
variable in the data frame. It is assumed the last variable in the data.}
}
\value{
A list with the percentage of inconsistent examples (pct) and a
nested list with the indexes of such examples (idx).
}
\description{
Compute the percentage of inconsistent examples, i.e., a set of examples
with the same attribute values and different class labels.
}
\examples{
model <- NCAR(data = qb, noise = 0.1)
df <- generateDF(model$parameters, 100)
df <- df |> dplyr::mutate(Error = NULL,
  Class = ObservedClass, ObservedClass = NULL)
result <- inconsistency(df)
result$pct
result$idx
}
