% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mux11}
\alias{mux11}
\title{11-input multiplexer data}
\format{
\subsection{\code{mux11}}{

A data frame with 10240 rows and 12 columns which :
\describe{
\item{A0}{Select line 0}
\item{A1}{Select line 1}
\item{A2}{Select line 2}
\item{D0}{Channel 0}
\item{D1}{Channel 1}
\item{D2}{Channel 2}
\item{D3}{Channel 3}
\item{D4}{Channel 4}
\item{D5}{Channel 5}
\item{D6}{Channel 6}
\item{D7}{Channel 7}
\item{Class}{Output}
}
}
}
\usage{
data(mux11)
}
\description{
Predict the output of a 11-input multiplexer.
Creator: D. Martínez-Galicia (davidgalicia@outlook.es).
Guided By : Dr. A. Guerra-Hernández (aguerra@uv.mx),
Dr. Francisco Grimaldo (francisco.grimaldo@uv.es),
Dr. Nicandro Cruz-Ramírez (ncruz@uv.mx),
Dr. Xavier Limón (hlimon@uv.mx)
Institution : Universidad Veracruzana, Mexico
and Universitat de València, Spain.
}
\keyword{datasets}
