% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{noiseLVL}
\alias{noiseLVL}
\title{Noise level}
\usage{
noiseLVL(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A float number.
}
\description{
Compute the noise level of data frame generated with the funtions of
noise models.
}
\examples{
data(qb)
model <- NCAR(data = qb, noise = 0.4)
df <- generateDF(model$parameters, 1000)
result <- noiseLVL(df)
result
}
