% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qb}
\alias{qb}
\title{Qualitative Bankruptcy Database}
\format{
\subsection{\code{qb}}{

A data frame with 250 rows and 7 columns:
\describe{
\item{IR}{Industrial Risk}
\item{MR}{Management Risk}
\item{FF}{Financial Flexibility}
\item{CR}{Credibility}
\item{CO}{Competitiveness}
\item{OP}{Operation Risk}
\item{Class}{Bankruptcy or Nonbankruptcy}
}
}
}
\source{
\url{https://doi.org/10.24432/C52889}
}
\usage{
data(qb)
}
\description{
Predict the Bankruptcy from Qualitative parameters from experts.
Creator: Mr. A. Martin (jayamartin@yahoo.com),
Mr. J. Uthayakumar (uthayakumar17691@gmail.com),
Mr. M. Nadarajan (nadaraj.muthuvel@gmail.com).
Guided By : Dr. V. Prasanna Venkatesan.
Institution : Sri Manakula Vinayagar Engineering College
and Pondicherry University, India,
}
\keyword{datasets}
